LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-intermediate-ignored-expected.txt


FAIL Starts rendering with flip-block assert_equals: expected 200 but got -100
FAIL No successful position (with intermediate successful), keep flip-block assert_equals: expected 250 but got -50
